+++ hypervisor/boot/include/guest/deprivilege_boot.h
@@ -19,6 +19,5 @@ const struct depri_boot_context *get_depri_boot_ctx(void);
 const struct lapic_regs *get_depri_boot_lapic_regs(void);
 
 struct vboot_operations* get_deprivilege_boot_ops(void);
-int32_t init_depri_vboot_info(__unused struct acrn_vm *vm);
 
 #endif /* end of include guard: DEPRIVILEGE_BOOT_H */
