+++ hypervisor/boot/include/guest/direct_boot.h
@@ -11,6 +11,5 @@
 #include <vboot.h>
 
 struct vboot_operations* get_direct_boot_ops(void);
-int32_t init_direct_vboot_info(struct acrn_vm *vm);
 
 #endif /* end of include guard: DIRECT_BOOT_H */
