+++ hypervisor/boot/include/guest/vboot.h
@@ -13,19 +13,16 @@ enum vboot_mode {
 	DEPRI_BOOT_MODE
 };
 
-struct acrn_vm;
 struct vboot_operations {
 	void (*init)(void);
 	uint64_t (*get_ap_trampoline)(void);
 	void *(*get_rsdp)(void);
 	void (*init_irq)(void);
-	int32_t (*init_vboot_info)(struct acrn_vm *vm);
 };
 
 void init_vboot_operations(void);
 void init_vboot(void);
 void init_vboot_irq(void);
-int32_t init_vm_boot_info(struct acrn_vm *vm);
 uint64_t get_ap_trampoline_buf(void);
 void *get_rsdp_ptr(void);
