+++ hypervisor/boot/include/guest/vboot_info.h
@@ -1,0 +1,13 @@
+/*
+ * Copyright (C) 2019 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef VBOOT_INFO_H
+
+#define VBOOT_INFO_H
+
+int32_t init_vm_boot_info(struct acrn_vm *vm);
+
+#endif /* end of include guard: VBOOT_INFO_H */
